feat(TDP-4285): First OS integ test based on api * Refactoring for EE integration.
