+++ dataprep-test-api/src/main/java/qa/spring/SpringContextTest.java
@@ -1,4 +1,4 @@
-package org.talend.dataprep.qa.spring;
+package qa.spring;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -8,7 +8,7 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.support.AnnotationConfigContextLoader;
 import org.talend.dataprep.helper.DataPrepAPIHelper;
-import org.talend.dataprep.qa.SpringContextConfiguration;
+import qa.SpringContextConfiguration;
 
 /**
  * This unit test is here to check that all bean is injected correctly
