+++ dataprep-test-api/src/test/java/org/talend/dataprep/OSRunnerConfigurationTest.java
@@ -1,4 +1,4 @@
-package org.talend.dataprep.qa;
+package org.talend.dataprep;
 
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
@@ -10,7 +10,7 @@ import org.junit.runner.RunWith;
 @RunWith(Cucumber.class)
 @CucumberOptions(plugin = {"pretty", "html:target/cucumber", "json:target/cucumber.json"},
         glue = "classpath:org/talend/dataprep/qa/step",
-        features = "classpath:features/os")
-public class OSRunnerConfigurationTest extends DataStepRunner {
+        features = "classpath:features")
+public class OSRunnerConfigurationTest {
 
 }
